+++ src/conversion_rates_contract.js
@@ -2,7 +2,7 @@ import Web3 from 'web3'
 import BigNumber from 'bignumber.js'
 
 import BaseContract from './base_contract.js'
-import conversionRatesABI from '../contracts/ConversionRatesContract.abi.json'
+import conversionRatesABI from '../abi/ConversionRatesContract.abi.json'
 import { validateAddress } from './validate.js'
 import { assertOperator, assertAdmin } from './permission_assert.js'
